requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
aiogram==3.2.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
websockets==12.0
aiohttp==3.9.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 